day3_factorio/requirements.txt
factorio-draftsman~=1.1.0
